+++ src/common/api/links.ts
@@ -258,10 +258,14 @@ export const links = {
       tasks: (id: ResourceId) => `/stacks/${stack}/builds/${id}/tasks`,
       log: (build: ResourceId) => `/stacks/${stack}/builds/${build}/log`,
     }),
-    hooks: (stack: ResourceId) => ({
-      collection: () => `/stacks/${stack}/hooks`,
-      single: (id: ResourceId) => `/stacks/${stack}/hooks/${id}`,
-      tasks: (id: ResourceId) => `/stacks/${stack}/hooks/${id}/tasks`,
+  }),
+
+  pipelines: () => ({
+    collection: () => `/pipelines`,
+    single: (id: ResourceId) => `/pipelines/${id}`,
+    keys: (pipeline: ResourceId) => ({
+      collection: () => `pipelines/${pipeline}/keys`,
+      single: (key: ResourceId) => `pipelines/${pipeline}/keys/${key}`,
     }),
   }),