+++ src/index.ts
@@ -23,6 +23,7 @@ import * as Structs from "./common/structs";
 import * as Notifications from "./notifications";
 import * as Websocket from "./common/api/websocket";
 import * as Surveys from "./resources/surveys";
+import * as Pipelines from "./resources/pipelines";
 
 export { Capability } from "./resources/hubs";
 
@@ -51,4 +52,5 @@ export {
   Structs,
   Websocket,
   Surveys,
+  Pipelines,
 };