+++ src/resources/environments/environment.ts
@@ -10,11 +10,10 @@ import {
   UserScope,
   StatefulCounts,
   UserIncludes,
-  IP,
   Cluster,
 } from "../../common/structs";
 import { ContainerState, Instances, Service } from "../containers";
-import { IPNet, Kind, IPState } from "../infrastructure/ips";
+import { IPNet } from "../infrastructure/ips";
 import { LoadBalancerService, VPNService, DiscoveryService } from "./services";
 import { Stack } from "../stacks";
 
@@ -47,6 +46,19 @@ export interface Environment extends Resource<EnvironmentMeta> {
   services: Services;
   private_network: PrivateNetwork | null;
 }
+
+export type NewEnvironment = {
+  name: string;
+  about: About;
+  cluster: Cluster;
+  stack: StackSummary | null;
+  features: Features;
+};
+
+type About = {
+  description: string;
+  favorite: boolean;
+};
 
 export interface PrivateNetwork {
   vxlan_tag: number;
@@ -81,11 +93,12 @@ export interface EnvironmentIncludes {
 }
 
 export interface EnvironmentMeta {
-  container_count?: StatefulCounts<ContainerState>;
+  containers_count?: StatefulCounts<ContainerState>;
   instances_count?: StatefulCounts<Instances.InstanceState>;
   containers?: {
     id: ResourceId;
     name: string;
+    cluster: string;
     state: State<ContainerState> & {
       desired: ContainerState;
     };