+++ src/resources/pipelines/index.ts
@@ -1,0 +1,2 @@
+export * as Tasks from "./tasks";
+export * from "./pipeline";