+++ CHANGELOG.md
@@ -9,7 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- `container.instances.delete` activity
+- `base_volume_gb` to server features to increase the base volume size
+
+* `container.instances.delete` activity
 
 ### Changed